% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eatMap.R
\name{eatMap}
\alias{eatMap}
\title{Plot the map of Germany}
\usage{
eatMap(
  data,
  config,
  lang = "de",
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{data}{Data frame. Data to be displayed.}

\item{config}{List. Configuration object for the display.}

\item{width}{Character. Width of the app.}

\item{height}{Character. Height of the app.}

\item{elementId}{Character. Id of the app.}
}
\description{
Plots the map of Germany.
}
\examples{
# eatMap()

}
